old/2025-jul-8/s6-ch-backtest/2025-jun-24/backend/requirements.txt
fastapi
httpx
clickhouse-connect
python-jose[cryptography]
python-dotenv
pydantic
